improve testing and add integration test for diff
